+++ R/methods.tolerance_delta.R
@@ -399,7 +399,7 @@ plot.tolerance_delta <- function(x,
                            paste0("Sympercent Difference between Methods (s%)")),
                     paste0("Difference between Methods (x - y)")),
          caption = cap1,
-         guides = "") +
+         dictionary = list(text = "")) +
     theme_bw() +
     theme(legend.position = "left",
           legend.title = element_blank())
@@ -434,7 +434,7 @@ plot.tolerance_delta <- function(x,
 #' @importFrom stats residuals lm na.omit pchisq shapiro.test ks.test rstudent df.residual anova rstandard sigma resid
 #' @export
 
-check.tolerance_delta <- function(x) {
+check.tolerance_delta <- function(x, ...) {
 
 
   call2 = x$call