+++ NEWS.md
@@ -1,5 +1,18 @@
 # NEWS
 
+# SimplyAgree 0.3.0
+
+- Add ehanced support for Deming regression.
+  - New function `deming_reli` for reliability analysis using Deming regression.
+  - Updated `dem_reg` function to include options for confidence intervals and plotting.
+- Added more power and sample size determination functions for limits of agreement.
+- Added Passing-Bablok regression function (`pb_reg`) for method comparison studies.
+- Updated methods for `simple_eiv` objects to be more extensive and act like other statistical models
+
+# SimplyAgree 0.2.2
+
+- Small fixes to errors related to plotting and df for lmer based results.
+
 # SimplyAgree 0.2.1
 
 - Add sympercent options for log transformed results in `tolerance_limit` and `agreement_limit`